Docs: Update `pydata-sphinx-theme` and add Discourse links
The `pydata-sphinx-theme` is updated to `v0.13.3` or higher, which is
the latest release. This changes the style a bit, but mostly for the
better.

It also allows to use custom icon links in the top-right header. This is
used to add a link to the AiiDA home page (using the AiiDA icon) and add
a link to the new Discourse server. Another admonition box is added to
the landing page that directs users needing support to Discourse.