Dependencies: Add new extra `tui` that provides `verdi` as a TUI (#6071)
The `tui` extra installs `trogon`. This package leverages `textual` to
turn `verdi`'s `click` interface into a Text-based User Interface (TUI).
It is added only if `trogon` is installed and can be imported. When it
is installed, it adds the `verdi tui` command, which launches the 
text-based interface of `verdi`.

Co-authored-by: Sebastiaan Huber <mail@sphuber.net>