Docker stack build CI action enabled for any changes
The docker build CI was only activated when `.docker` or
`.github/workflows/docker*.yaml` are changed. It prevent the build
check from changes in src code which may potentially break the
docker stack build. For example when CLI interface for `quicksetup`
is changed or the interface for `computer/code` setup is
changed, since the docker stack builds relies on running these commands inside.

We use paths-ignore to ignore the trigger of docker stack build when changes
are from `docs/` or `tests/`.