CLI: Fix `repository` being required for `verdi quicksetup` (#6129)
Regression added by c53ea20a497f66bc88f68d0603cf9a32614fc4c2 which made
the `--repository` option for `verdi setup` required, as it should be.
However, it did so by making the base option required. The problem is
that the option for both `verdi setup` as well as `verdi quicksetup`
inherit from this, but for `verdi quicksetup` it should not be required
as the default will be populated automatically. As an alternative, the
option specific for `verdi setup` is now made required.