Tests: Fix failing `tests/cmdline/commands/test_setup.py`
The previous commit fixed a bug in the evaluation of defaults for
various options of the `verdi setup` command. Due to the bug, these
options would set a default even if the corresponding config option was
not defined. Instead of no default being defined, the empty tuple `()`
would be set as string value.

As soon as the bug was fixed, the `test_setup_profile_uuid` test started
failing since it doesn't explicitly defined values for the options
`--email`, `--first-name`, `--last-name` and `--institution`.