Add the `SshFabricTransport` plugin
This transport plugin subclasses the `SshTransport` plugin but replaces
the managing of the connection with the `fabric` library. Under the hood
this library still uses `paramiko`, however, it is a lot clevered in
automatically determining the correct connection configuration. For
example, it will automatically search for SSH configurations on the
machine. This makes it easier for the user to setup the transport to
connect to a particular host if they can connect to it through a direct
`ssh` invocation`. Since `fabric` still uses `paramiko` under the hood,
just like the `SshTransport`, most of the implementation can be reused.