CLI: Do not load config in defaults and callbacks during tab-completion (#6144)
The `get_default_profile` default of the `PROFILE` option and the
`set_log_level` callback of the `VERBOSITY` option both load the config.
Since defaults and callbacks are also evaluated during tab-completion
this was slowing down tab-completion significantly since loading the
config has a non-negligible cost.

The `set_log_level` callback is modified to explicitly check whether
we are are tab-completing, in which case `ctx.resilient_parsing` is set
to `True`. In this case, the functions now returns `None` and no longer
loads the config.

For `get_default_profile`, the `CallableDefaultOption` class is added
which allows the default to be made a callable, which will return `None`
if `ctx.resilient_parsing` is set to `True`.