`FolderData`: Expose repository API on top-level namespace (#6150)
In 8293e453789d0bad9cf631ecfc08542dd9ad892d, the `Node` interface was
refactored to move the API of the repository to the `base.repository`
namespace. The original methods would be forwarded with a deprecation
message being printed.

Although this made sense for most node types, in an effort to clean up
the node interface which was overpopulated, for the `FolderData` the
repository interface is the main interface and it doesn't make sense to
force the users to go all the way down to the nested `base.repository`
namespace to access it. Therefore the public API of the repository is
restored on the top-level namespace of the `FolderData` class.