CLI: Add `cached` and `cached_from` projections to `verdi process list`
The `cached` projection will print a checkmark for nodes that were
cached from another node. The `cached_from` projection will show the
UUID of the cache source, if it exists.

The `cached` projection is added to the default projections for `verdi
process list`. The use of caching is becoming more prevalent, but often
users can still be surprised by certain behavior when processes are
taken from cache when they didn't expect it. By showing whether a
process was taken from cache or not by default in the output of `verdi
process list` should provide clarity since this is often the first place
that users check.