Tests: Move memory leak tests to main unit test suite
The `.github/system_tests/pytest/test_memory_leaks.py` file provided
tests to ensure memory is not being leaked when running processes. These
tests do not require being executed in standalone `pytest` invocation
but can be included in the main unit test suite. Historically, the
separation was required when the main unit test suite was not fully
using `pytest` yet but used a framework based on `unittest`.

With this migration, the last test in the `.github/workflows/tests.sh`
script has been moved and now it merely calls the main test suite. The
CI workflows that called it, now simply directly invoke the command to
run the main test suite and the `tests.sh` script is deleted.