CLI: Fix `verdi config set` when setting list option (#6166)
`verdi config set` would except when setting a single value for an
option that is of list type, such as `caching.enable_for`. This only
started happening after the recent move to `pydantic` for the
configuration options. Now the `Option.validate` will correctly raise
when trying to set a string value for a list type.

The `verdi config set` implementation is updated to check when it is
setting a value for an option with a list type, and in that case, the
value is wrapped in a list, unless the `--append` or `--remove` flags
are specified.