Docker: Add folders that automatically run scripts before/after daemon start (#6170)
In order to simplify the implementation of using the `aiida-core` image
as the base for customized images, the `run-before-daemon-start` and
`run-after-daemon-start` script folders are created. Any executables in
these two folders will be executed before and after the AiiDA daemon is
started in the container, respectively.

The standard linux `run-parts` tool is used to scan these folders for
files, which are run in the lexical sort order of their names, according
to the C/POSIX locale character collation rules

Cherry-pick: 5fe6f4fda1eaf5a1e05e87063aa4eafa47fa1c07