Add the `report` method to `logging.LoggerAdapter` (#6186)
AiiDA defines the `REPORT` log level and adds the `report` method to the
`logging.Logger` class so a log message can easily be emitted at that
level. However, the logger of `Process` instances is a `LoggerAdapter`
which does not inherit from `Logger` so the method also needs to be
added there independently. Without this fix, calling `self.logger.report` in
the `Parser.parse` method would raise an `AttributeError`.