Engine: Refactor handling of `remote_folder` and `retrieved` outputs
The handling of outputs of a process are handled as follows. A new
output is registered by calling `Process.out`. This validates the output
against the process specification, and if it passes, the output is added
to the `Process._outputs` mapping. Whenever a new state is entered,
`Process.on_entered` is triggered which calls `Process.update_outputs`.
This goes through all the outputs in memory and adds a link to the
process node if it hasn't already been done before.

The `CalcJob` implementation wen around this mechanism though for the
`remote_folder` and `retrieved` outputs. These outputs are created in
the `upload_calculation` and `retrieve_calculation` methods of the
`aiida.engine.daemon.execmanager`, respectively, and those same methods
also create the link to the process node.

To restore consistency, the `upload_calculation`/`retrieve_calculation`
methods now return the node they created and it is the caller, the
`task_upload_job` and `task_retrieve_job`, respectively, that now call
`Process.out` for the returned node. At the end of those tasks, the
process will change to a new state and the outputs will be "flushed" to
the database by `update_outputs`.

This change now also allows removing the "manual" handling of outputs
in the `CalcJob.parse` call.