CLI: Add the `verdi process repair` command
This command replaces `verdi devel rabbitmq tasks analyze`. This command
was added to the `verdi devel` namespace because it is working around a
problem and it was experimental. Since then, it has proved really
efficient and so should be made more directly available to users in case
of stuck processes.

The implementation is moved to `verdi process repair` and the original
command simply forwards to it, while emitting a message that it is
deprecated.

While the original command would not do anything by default and the
`--fix` flag had to be explicitly specified, this behavior is inverted
for `verdi process repair`. By default it will fix inconsistencies and
the `--dry-run` flag can be used to have to old behavior of just
detecting them.