Add the `aiida.common.log.capture_logging` utility
The `capture_logging` is a context manager that yields a stream in
memory to which all content written to the specified logger is
duplicated. This does not interfere with any existing logging handlers
whatsoever and so is non-destructive. It is useful to capture any output
that is logged into memory in order to be able to act on it.