Add support for `NodeLinksManager` to YAML serializer (#6199)
The `Node.inputs` and `Node.outputs` properties return instances of the
`aiida.orm.utils.managers.NodeLinksManager` class. Support is added to
the `aiida.orm.utils.serialize` YAML serializers such that these
instances can now be stored in the context of `WorkChains` as these are
serialized to YAML for the checkpoints.