Process functions: Fix bug with variable arguments (#6201)
The process function implementation contained a bug where a function
that specified variable positional arguments followed by keyword
arguments would not be accepted. For example:

    def function(*args, arg_a, arg_b):
        pass

    function(*(1, 2), 3, 4)

is a perfectly valid function definition and call but it would not work
when decorated into a process function.

Part of the problem was that the class argument `_varargs` of the
dynamically constructed `FunctionProcess` was used for the name of
variable positional as well as keyword arguments. If both were defined,
the former would be overridden by the latter. This is now split in
`_var_positional` and `_var_keyword` respectively.

The conversion of the original positional and keyword arguments passed
to the function into the process input dictionary is simplified. As well
as for the reverse process where the process inputs are converted back
in to positional and keyword arguments before passing them to the
wrapped function.