CLI: Fix `DynamicEntryPointCommandGroup` for default values
Options generated for model fields that do not define a default would
get the `PydanticUndefined` value as default, because this is the value
that pydantic's `FieldInfo.default` returns when no explicit default was
set. This is to be able to distinguish it from `None` as a default value.

The `default_factory` attribute was also not respected. The problem is
fixed by explicitly checking for `PydanticUndefined` as the value of the
`FieldInfo.default` property. If that is the case, the value is set to
`FieldInfo.default_factory` which returns `None` if not set. The
`Option.default` attribute of `click` takes a callable, so if the
`default_factory` is defined, this should work.