Fixtures: Fix `suppress_warnings` of `run_cli_command`
Now that the `warnings.showdeprecations` config option is once again
respected by `aiida.common.warnings.warn_deprecation`, the
`run_cli_command` needs to explicitly set it to `False` for the duration
of the command if `suppress_warnings` is set to `True`.

In addition, if `use_subprocess` is `True`, the config needs to be
stored to disk such that the change is picked up by the command being
executed in the subprocess.