Config: Fix `RecursionError` in `_merge_deprecated_cache_yaml`
The `_merge_deprecated_cache_yaml` function logs a deprecation warning
when the caching config is still in the old stand-alone format. This
started raising a `RecursionError` because the `warn_deprecation`
utility tries to load the `warnings.showdeprecations` option, which
requires loading the config, which hits `_merge_deprecated_cache_yaml`.

The workaround is to not use the `warn_deprecation` utility but directly
use the `warnings.warn` method of the standard library. It makes sense
to always show this deprecation warning as it should only be shown once
anyway, the first time that the old config is migrated.